feat: mechanism to spawn dead enemies corpses
