--- Erlang-Legacy/Assets/Core/Utility/DeadBodiesManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ba4dfa62a59d72f4b9c17709c222e5be
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
